fix(logging): gke resource autodetection
